requirements.txt
sentence-transformers
langchain-huggingface
langchain
flask
pypdf
python-dotenv
pinecone
langchain-pinecone
langchain_community
langchain_openai
langchain_experimental
google-generativeai
networkx
matplotlib
langchain_google_genai
flask-session
-e .
